requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
pytz
sqlparse
tzdata
whitenoise
python-decouple
psycopg2-binary
Pillow
django-cloudinary-storage
cloudinary
qrcode[pil]
pusher
django-jazzmin
django-rest-swagger
drf-yasg